Cypress updates, Refactor Edit, Open Status bug
