+++ frontend/cypress/support/index.ts
@@ -9,6 +9,7 @@ declare namespace Cypress {
     typeAndTriggerChange(value: string): Chainable<void>;
     navigateToDetailsScreen(complaintType: string, complaintIdentifier: string) : Chainable<any>;
     navigateToEditScreen(complaintType: string, complaintIdentifier: string) : Chainable<any>;
+    navigateToCreateScreen() : Chainable<any>;
     isInViewport() : Chainable<any>;
     waitForSpinner(): Chainable<any>;
     clearFilterById(filterId: string): Chainable<any>;
